requirements.txt
crispy-bootstrap4
Django~=4.0
django-bootstrap4
django-crispy-forms
django-import-export
django-magiclink
django-tables2
gunicorn
ovh~=1.0.0
pytest-bdd~=5.0.0
python-decouple
tablib~=3.3
whitenoise
django-phonenumber-field[phonenumbers]~=7.0.2
